Fix small PR comments
